# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI-Powered School Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /schools/{schoolId}/classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

  # ATTENDANCE
  /schools/{schoolId}/attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: Get attendance records
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for students
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "200":
          description: Attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /schools/{schoolId}/attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary stats
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # FEES
...
  /schools/{schoolId}/announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: audience
          in: query
          schema:
            type: string
            enum: [all, teachers, students, parents]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /schools/{schoolId}/announcements/{announcementId}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: announcementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, school_admin, principal, teacher, accountant, parent, student]
        schoolId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # SCHOOLS
...
    Class:
      type: object
      required: [id, name, schoolId, studentCount]
      properties:
        id:
          type: integer
        name:
          type: string
        section:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        classTeacherId:
          type: ["integer", "null"]
        classTeacherName:
          type: ["string", "null"]
        schoolId:
          type: integer
        studentCount:
          type: integer
        capacity:
          type: ["integer", "null"]
    ClassInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        section:
          type: string
        grade:
          type: integer
        classTeacherId:
          type: integer
        capacity:
          type: integer

    # ATTENDANCE
    AttendanceRecord:
      type: object
      required: [id, studentId, date, status, schoolId]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        admissionNumber:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late, excused]
        note:
          type: ["string", "null"]
        schoolId:
          type: integer
    AttendanceInput:
      type: object
      required: [date, classId, records]
      properties:
        date:
          type: string
          format: date
        classId:
          type: integer
        records:
          type: array
          items:
            type: object
            required: [studentId, status]
            properties:
              studentId:
                type: integer
              status:
                type: string
                enum: [present, absent, late, excused]
              note:
                type: string
    AttendanceSummary:
      type: object
      required: [totalStudents, avgAttendanceRate, presentToday, absentToday, lateToday, byClass]
      properties:
        totalStudents:
          type: integer
        avgAttendanceRate:
          type: number
        presentToday:
          type: integer
        absentToday:
          type: integer
        lateToday:
          type: integer
        byClass:
          type: array
          items:
            type: object
            required: [classId, className, rate]
            properties:
              classId:
                type: integer
              className:
                type: string
              rate:
                type: number

    # FEES
...
    Announcement:
      type: object
      required: [id, title, content, audience, schoolId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        audience:
          type: string
          enum: [all, teachers, students, parents]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        schoolId:
          type: integer
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, content, audience]
      properties:
        title:
          type: string
        content:
          type: string
        audience:
          type: string
          enum: [all, teachers, students, parents]
        priority:
          type: string
          enum: [low, normal, high, urgent]

    # DASHBOARD
...