# lib/db/src/schema/announcements.ts
import { pgTable, text, serial, timestamp, integer, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { schoolsTable } from "./schools";
import { usersTable } from "./users";

export const audienceEnum = pgEnum("audience", ["all", "teachers", "students", "parents"]);
export const priorityEnum = pgEnum("priority", ["low", "normal", "high", "urgent"]);

export const announcementsTable = pgTable("announcements", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  content: text("content").notNull(),
  audience: audienceEnum("audience").notNull().default("all"),
  priority: priorityEnum("priority").notNull().default("normal"),
  schoolId: integer("school_id").notNull().references(() => schoolsTable.id),
  authorId: integer("author_id").references(() => usersTable.id),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertAnnouncementSchema = createInsertSchema(announcementsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertAnnouncement = z.infer<typeof insertAnnouncementSchema>;
export type Announcement = typeof announcementsTable.$inferSelect;